refactor: move file functions
